chore(EMS-1491): rename 'exporterCompany' to 'company'
