+++ src/api/custom-resolvers/mutations/index.ts
@@ -8,7 +8,7 @@ import addAndGetOTP from './add-and-get-OTP';
 import sendEmailPasswordResetLink from './send-email-password-reset-link';
 import accountPasswordReset from './account-password-reset';
 import deleteApplicationByReferenceNumber from './delete-application-by-refrence-number';
-import updateCompanyAndExporterCompanyAddress from './update-exporter-company-and-company-address';
+import updateCompanyAndExporterCompanyAddress from './update-company-and-company-address';
 import submitApplication from './submit-application';
 import sendEmailInsuranceFeedback from './send-email-insurance-feedback';
