+++ src/api/schema.ts
@@ -98,7 +98,7 @@ export const lists = {
             };
 
             // generate a new `exporter company address` relationship with the company
-            await context.db.ExporterExporterCompanyAddress.createOne({
+            await context.db.ExporterCompanyAddress.createOne({
               data: {
                 company: {
                   connect: {
